+++ class/shippableorder.class.php
@@ -297,8 +297,8 @@ class ShippableOrder
 			return '';
 		}
 
-        $picto = self::getPicto($isShippable['to_ship'], $isShippable['shippable'], $isShippable['stock'], $isShippable['qty_shippable'], $line);
-        $virtualPicto = self::getPicto($isShippable['to_ship'], $isShippable['isShippableVirtual'], $isShippable['stock_virtuel'], $isShippable['qtyShippableVirtual'], $line);
+        $picto = self::getPicto($isShippable['to_ship'], $isShippable['shippable'], $isShippable['stock'], $isShippable['qty_shippable'], $line, 'stock');
+        $virtualPicto = self::getPicto($isShippable['to_ship'], $isShippable['isShippableVirtual'], $isShippable['stock_virtuel'], $isShippable['qtyShippableVirtual'], $line, 'virtualStock');
 
 		if($withStockVisu) {
 			return array($isShippable['stock'].' '.$picto, $isShippable['stock_virtuel'].' '.$virtualPicto);
@@ -319,16 +319,16 @@ class ShippableOrder
         $out = $langs->trans('VirtualStockDetailHeader');
 
 		if (!empty($conf->mrp->enabled)) {
-			$out .= $langs->trans('VirtualStockDetail', 'ordres de fabrication','"Validé" et "En cours"');
+			$out .= $langs->trans('VirtualStockDetail', $langs->trans('MO'), $langs->trans('MO_status'));
 		}
 
 		// Stock decrease mode
 		if (!empty($conf->global->STOCK_CALCULATE_ON_SHIPMENT) || !empty($conf->global->STOCK_CALCULATE_ON_SHIPMENT_CLOSE) || !empty($conf->global->STOCK_CALCULATE_ON_BILL)) {
-            $out .= $langs->trans('VirtualStockDetail', 'commandes','"Validé" et "En cours"');
+            $out .= $langs->trans('VirtualStockDetail', $langs->trans('Orders'), $langs->trans('Orders_status'));
 			if (!empty($conf->global->STOCK_CALCULATE_ON_SHIPMENT)) {
-				$out .= $langs->trans('VirtualStockDetail', 'expeditions','"Validé" et "Cloturé"');
+				$out .= $langs->trans('VirtualStockDetail', $langs->trans('Shipments'), $langs->trans('Shipments_status'));
 			} elseif (!empty($conf->global->STOCK_CALCULATE_ON_SHIPMENT_CLOSE)) {
-				$out .= $langs->trans('VirtualStockDetail', 'expeditions','"Cloturé"');
+				$out .= $langs->trans('VirtualStockDetail', $langs->trans('Shipments'), $langs->trans('Shipments_status_closed'));
 			}
 
 		}
@@ -338,12 +338,12 @@ class ShippableOrder
             || !empty($conf->global->STOCK_CALCULATE_ON_SUPPLIER_DISPATCH_ORDER)
             || !empty($conf->global->STOCK_CALCULATE_ON_SUPPLIER_VALIDATE_ORDER)
             || !empty($conf->global->STOCK_CALCULATE_ON_SUPPLIER_BILL)) {
-            $statusCmdFourn = '"Accepté", "Envoyé", "Reçu partiellement"';
+            $statusCmdFourn = 'PO_status';
             if (isset($includedraftpoforvirtual)) {
-                $statusCmdFourn .= ',"Brouillon", "Validé"';
+                $statusCmdFourn .= '_all';
             }
-            if(empty($conf->global->STOCK_CALCULATE_ON_SUPPLIER_VALIDATE_ORDER)) $out .= $langs->trans('VirtualStockDetail', 'commandes fournisseurs',$statusCmdFourn);
-            $out .= $langs->trans('VirtualStockDetailReception');
+            if(empty($conf->global->STOCK_CALCULATE_ON_SUPPLIER_VALIDATE_ORDER)) $out .= $langs->trans('VirtualStockDetail', $langs->trans('PO'), $langs->trans($statusCmdFourn));
+            $out .= $langs->trans('VirtualStockDetail', $langs->trans('Receptions'), $langs->trans('Receptions_status'));
 
 		}
         return $out;
@@ -351,16 +351,17 @@ class ShippableOrder
 
     /**
      * Get picto shippable order line
-     * @param float $toship
+     * @param string $type
+	 * @param float $toship
      * @param float $shippable
      * @param float $stock
      * @param float $qty_shippable
      * @param OrderLine $line
      * @return string
      */
-    public static function getPicto($toship, $shippable, $stock, $qty_shippable, $line) {
+    public static function getPicto($toship, $shippable, $stock, $qty_shippable, $line, $type = 'stock') {
         $pictopath = self::getPictoPath($toship, $shippable);
-		$infos = self::getPictoInfos($stock, $toship, $qty_shippable);
+		$infos = self::getPictoInfos($stock, $toship, $qty_shippable, $type);
         $picto = '<img src="'.$pictopath.'" border="0" title="'.$infos.'">';
 		if($toship > 0 && $toship != $line->qty) {
 			$picto.= ' ('.$toship.')';
@@ -370,17 +371,24 @@ class ShippableOrder
 
     /**
      * Get picto qty infos
-     * @param float $stock
+     * @param string $type
+	 * @param float $stock
      * @param float $toship
      * @param float $qty_shippable
      * @return string
      */
-    public static function getPictoInfos($stock, $toship, $qty_shippable) {
+    public static function getPictoInfos($stock, $toship, $qty_shippable, $type = 'stock') {
         global $langs;
-        $infos = $langs->trans('QtyInStock', $stock);
-        $infos .= " - ".$langs->trans('RemainToShip', $toship);
-        $infos .= " - ".$langs->trans('QtyShippable', $qty_shippable);
-        return $infos;
+        if($type == 'stock') {
+			$infos = $langs->trans('QtyInStock', $stock);
+			$infos .= " - ".$langs->trans('RemainToShip', $toship);
+			$infos .= " - ".$langs->trans('QtyShippable', $qty_shippable);
+			return $infos;
+		} else if($type == 'virtualStock') {
+			$infos = $langs->trans('VirtualQtyInStock', $stock);
+			$infos .= " - ".$langs->trans('RemainToShip', $toship);
+			return $infos;
+		}
     }
 
     /**
@@ -673,12 +681,12 @@ class ShippableOrder
 
 			
 			if($nbShippingCreated > 0) {
-				if($conf->global->SHIPPABLEORDER_GENERATE_GLOBAL_PDF) $this->generate_global_pdf($TFiles);	
+				if(!empty($conf->global->SHIPPABLEORDER_GENERATE_GLOBAL_PDF)) $this->generate_global_pdf($TFiles);
 				
 				setEventMessage($langs->trans('NbShippingCreated', $nbShippingCreated));
 				$dol_version = (float) DOL_VERSION;
 				
-				if ($conf->global->SHIPPABLE_ORDER_DISABLE_AUTO_REDIRECT)
+				if (!empty($conf->global->SHIPPABLE_ORDER_DISABLE_AUTO_REDIRECT))
 				{
 
 					header("Location: ".$_SERVER["PHP_SELF"].'?'.http_build_query($TURL) );					
@@ -692,7 +700,7 @@ class ShippableOrder
 				setEventMessage($langs->trans('NoOrderSelectedOrAlreadySent'), 'warnings');
 				$dol_version = (float) DOL_VERSION;
 				
-				if ($conf->global->SHIPPABLE_ORDER_DISABLE_AUTO_REDIRECT)
+				if (!empty($conf->global->SHIPPABLE_ORDER_DISABLE_AUTO_REDIRECT))
 				{
 					header("Location: ".$_SERVER["PHP_SELF"].'?'.http_build_query($TURL) );					
 				}else{